Remove pkf
