--- docs/source/reference/nn/functional.rst
@@ -1,4 +1,0 @@
-Functional
-==========
-.. automodapi:: pykeen.nn.functional
-    :no-heading:
