+++ src/pykeen/utils.py
@@ -1644,3 +1644,29 @@ if __name__ == "__main__":
     import doctest
 
     doctest.testmod()
+
+
+def circular_correlation(a: FloatTensor, b: FloatTensor) -> FloatTensor:
+    """
+    Compute the circular correlation between to vectors.
+
+    .. note ::
+        The implementation uses FFT.
+
+    :param a: shape: s_1
+        The tensor with the first vectors.
+    :param b:
+        The tensor with the second vectors.
+
+    :return:
+        The circular correlation between the vectors.
+    """
+    # Circular correlation of entity embeddings
+    a_fft = torch.fft.rfft(a, dim=-1)
+    b_fft = torch.fft.rfft(b, dim=-1)
+    # complex conjugate
+    a_fft = torch.conj(a_fft)
+    # Hadamard product in frequency domain
+    p_fft = a_fft * b_fft
+    # inverse real FFT
+    return torch.fft.irfft(p_fft, n=a.shape[-1], dim=-1)
